SONIC-1568: fix dialog dimensions
