+++ Backpack/src/main/res/layout/bpk_dialog.xml
@@ -1,89 +1,75 @@
 <?xml version="1.0" encoding="utf-8"?>
-<net.skyscanner.backpack.dialog.internal.DialogInsetsLayout xmlns:android="http://schemas.android.com/apk/res/android"
+
+<androidx.appcompat.widget.LinearLayoutCompat xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:background="@color/bpkTextSecondary">
+    android:layout_height="wrap_content"
+    android:orientation="vertical"
+    app:layout_constraintBottom_toBottomOf="parent"
+    app:layout_constraintEnd_toEndOf="parent"
+    app:layout_constraintStart_toStartOf="parent"
+    app:layout_constraintTop_toTopOf="parent">
 
-    <net.skyscanner.backpack.dialog.internal.DialogWindowLayout
-        android:id="@+id/dialog_window_layout"
+    <ScrollView
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:fadingEdge="vertical"
+        android:fadingEdgeLength="@dimen/bpkSpacingLg"
+        android:requiresFadingEdge="vertical"
+        android:scrollbarStyle="outsideOverlay">
 
         <androidx.appcompat.widget.LinearLayoutCompat
-            android:layout_width="0dp"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:orientation="vertical"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintWidth_max="@dimen/bpk_dialog_max_width">
+            android:background="@drawable/bpk_dialog_bg_top"
+            android:orientation="vertical">
 
-            <ScrollView
+            <net.skyscanner.backpack.dialog.internal.BpkDialogIcon
+                android:id="@+id/dialog_icon"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center_horizontal"
+                tools:src="@drawable/bpk_trash" />
+
+            <net.skyscanner.backpack.text.BpkText
+                android:id="@+id/dialog_title"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:fadingEdge="vertical"
-                android:fadingEdgeLength="@dimen/bpkSpacingLg"
-                android:requiresFadingEdge="vertical"
-                android:scrollbarStyle="outsideOverlay">
+                android:layout_marginStart="@dimen/bpkSpacingLg"
+                android:layout_marginTop="@dimen/bpkSpacingLg"
+                android:layout_marginEnd="@dimen/bpkSpacingLg"
+                android:textAlignment="center"
+                android:textColor="@color/bpkTextPrimary"
+                app:textStyle="heading3"
+                tools:text="Title" />
 
-                <androidx.appcompat.widget.LinearLayoutCompat
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:background="@drawable/bpk_dialog_bg_top"
-                    android:orientation="vertical">
-
-                    <net.skyscanner.backpack.dialog.internal.BpkDialogIcon
-                        android:id="@+id/dialog_icon"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_gravity="center_horizontal"
-                        tools:src="@drawable/bpk_trash" />
-
-                    <net.skyscanner.backpack.text.BpkText
-                        android:id="@+id/dialog_title"
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:layout_marginStart="@dimen/bpkSpacingMd"
-                        android:layout_marginTop="@dimen/bpk_dialog_spacing_title"
-                        android:layout_marginEnd="@dimen/bpkSpacingMd"
-                        android:textAlignment="center"
-                        android:textColor="@color/bpkTextPrimary"
-                        app:textStyle="heading3"
-                        tools:text="Title" />
-
-                    <net.skyscanner.backpack.text.BpkText
-                        android:id="@+id/dialog_description"
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:layout_marginStart="@dimen/bpkSpacingMd"
-                        android:layout_marginTop="@dimen/bpkSpacingBase"
-                        android:layout_marginEnd="@dimen/bpkSpacingMd"
-                        android:layout_marginBottom="@dimen/bpkSpacingLg"
-                        android:textAlignment="center"
-                        android:textColor="@color/bpkTextPrimary"
-                        app:textStyle="bodyDefault"
-                        tools:text="Message" />
-
-                </androidx.appcompat.widget.LinearLayoutCompat>
-
-            </ScrollView>
-
-            <androidx.appcompat.widget.LinearLayoutCompat
-                android:id="@+id/dialog_buttons_root"
+            <net.skyscanner.backpack.text.BpkText
+                android:id="@+id/dialog_description"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:background="@drawable/bpk_dialog_bg_bottom"
-                android:orientation="vertical"
-                android:paddingBottom="@dimen/bpk_dialog_padding_bottom"
-                app:divider="@drawable/bpk_dialog_bg_divider"
-                app:showDividers="middle|end" />
+                android:layout_marginStart="@dimen/bpkSpacingLg"
+                android:layout_marginTop="@dimen/bpkSpacingMd"
+                android:layout_marginEnd="@dimen/bpkSpacingLg"
+                android:textAlignment="center"
+                android:textColor="@color/bpkTextPrimary"
+                app:textStyle="bodyDefault"
+                tools:text="Message" />
 
         </androidx.appcompat.widget.LinearLayoutCompat>
 
-    </net.skyscanner.backpack.dialog.internal.DialogWindowLayout>
+    </ScrollView>
 
-</net.skyscanner.backpack.dialog.internal.DialogInsetsLayout>
+    <androidx.appcompat.widget.LinearLayoutCompat
+        android:id="@+id/dialog_buttons_root"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical"
+        android:background="@drawable/bpk_dialog_bg_bottom"
+        android:paddingTop="@dimen/bpk_dialog_padding_bottom"
+        android:paddingBottom="@dimen/bpk_dialog_padding_bottom"
+        app:divider="@drawable/bpk_dialog_bg_divider"
+        app:showDividers="middle" />
+
+</androidx.appcompat.widget.LinearLayoutCompat>
