--- Backpack/src/main/res/values-sw600dp/dimens.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <!-- internal resources -->
-    <dimen name="bpk_dialog_max_width">580dp</dimen>
-</resources>
